Improve sync times
This is a bit uglier than I would like, but unfortunately we can't ship
off the async task join to a tokio task as
we spawn so many tasks that tokio won't schedule it. So instead we spawn
the rayon scope on a separate thread first. This cuts sync times by
~25+%